updating .gitignore and adding some comments.
Tried to abstract some of the err = sprintf(...), if err < 0
stuff away and it messed with the determinism and added jitter
for reasons I don't want to try to think about or reason about.

Maybe there is a better way of doing it, but for now, leaving
stuff as is.